backend/requirements.txt
psycopg2-binary
python-dotenv
langchain
tiktoken
chromadb
openai
fastapi
uvicorn[standard]
pydantic
pandas
tqdm
tenacity
numpy